// workflow_fe/src/lib/api/gen/types/WorkflowActionResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { WorkflowUserResponseDto } from "./WorkflowUserResponseDto.ts";

export type WorkflowActionResponseDtoActionEnum =
  | "TRIGGERED"
  | "STEP_ACTIVATED"
  | "APPROVED"
  | "REJECTED"
  | "COMMENTED"
  | "REASSIGNED"
  | "CANCELLED"
  | "FAILED";

export type WorkflowActionResponseDto = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  workflowInstanceId: string;
  /**
   * @type string
   */
  workflowStepId: string | null;
  /**
   * @type string
   */
  action: WorkflowActionResponseDtoActionEnum;
  /**
   * @type string
   */
  actorUserId: string | null;
  /**
   * @type object
   */
  actorUser: WorkflowUserResponseDto | null;
  /**
   * @type string
   */
  comment: string | null;
  /**
   * @type string
   */
  reason: string | null;
  /**
   * @type object
   */
  metadataJson: object | null;
  /**
   * @type string
   */
  createdAt: string;
};