// workflow_fe/src/lib/api/gen/types/WorkflowApprovalRuleResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { WorkflowApprovalStepConfigResponseDto } from "./WorkflowApprovalStepConfigResponseDto.ts";

export type WorkflowApprovalRuleResponseDto = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  workflowTemplateId: string;
  /**
   * @type string
   */
  name: string;
  /**
   * @type number
   */
  priority: number;
  /**
   * @type object
   */
  conditionJson: object | null;
  /**
   * @type boolean
   */
  isFallback: boolean;
  /**
   * @type boolean
   */
  isActive: boolean;
  /**
   * @type array
   */
  steps: WorkflowApprovalStepConfigResponseDto[];
  /**
   * @type string
   */
  createdAt: string;
  /**
   * @type string
   */
  updatedAt: string;
};