// src/lib/api/gen/types/RbacPermissionResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type RbacPermissionResponseDto = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  name: string;
  /**
   * @type string
   */
  slug: string;
  /**
   * @type string
   */
  description: string | null;
  /**
   * @type string
   */
  resource: string;
  /**
   * @type string
   */
  action: string;
  /**
   * @type string
   */
  createdAt: string;
  /**
   * @type string
   */
  updatedAt: string;
};